// frontend_native/api/generated/model/signupRequest.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Awesome Template API
 * Awesome Template バックエンドサービスの API ドキュメントです。
 * OpenAPI spec version: 1.0.0
 */

/**
 * 新しいユーザーを登録するためのリクエストペイロードです
 */
export interface SignupRequest {
  /** ユーザーのメールアドレス */
  email?: string;
  /** ユーザーの表示名 */
  name?: string;
}